requirements.txt
# Backend зависимости
fastapi
uvicorn[standard]

tortoise-orm
aiosqlite
aiofiles
httpx
python-multipart
wotreplay

authlib
python-dotenv
PyJWT